requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
python-dotenv
httpx>=0.27
python-dotenv>=1.0

